math/xy_series.py: Provide a better get_avg_value() method..
The previous get_avg_value() implementation was only considering data
points within the (x0, x1) bounds and not interpolating the data out to
x0 and x1.  This creates unused gaps in the data when resampling, for
instance, and it doesn't provide as good of an average as we could get
if we used the full range of data from x0 to x1.

We first provide _integrate() to do generic integration and then
reimplement get_avg_value() in terms of the new _integrate() method.  A
short test sequence is also provided to ensure correctness.