p_sensor/gl_track_mode.py: Deal with Windows minimization better.
On Windows, when we get minimized we stop getting refresh events from
glfw.  This would cause points to pile up in the queue (which is fine)
and then when we expand the window again the processing of all the
queued points didn't work properly.  This is all sorted out now.