p_sensor/__init__.py: Provide make_xti() which specifically makes an XTI sensor.
Soon we will introduce support for the XMHTI sensor which has some
similar method names to XTI but different argument counts.  This makes
the linters unhappy since the simple make() function could return either
an XTI or and XMHTI and then it got confused when you used on of these
similarly-named methods with the "wrong" number of arguments.