z_sensor/tcsc.py: Add support for TCSC 1.0.8 to rotation Z before peak-fitting.
When we try to sweep higher-frequency crystals such as the 262 kHz
temperature crystal, we get a lot of additional phase-shift in the
measurement due to the lowish bandwidth of the sense op-amp.  This
distorts Real(Z) enough that the firmware can no longer fit the peak.
By applying a rotation to Z in the Nyquist plane before fitting the
data, we can compensate for the phase shift and get a clean peak.