z_sensor/tcsc.py: Allow applying a phase shift to the impedance data when reading it from the sensor.
In the previous commit, we added support for the sensor firmware to
phase-shift the impedance data before curve-fitting it.  However, the
sensor still returns the raw, unshifted data to us when we issue a read
command.  This commit allows us to apply a phase shift to the data after
we read it, with the intention that it could be shifted to match the
same shift that was used by the firmware during curve-fitting.