z_sensor/peak_tracker.py: If the firmware fit uses the maximum number of iterations, don't consider it a fit failure.
Sometimes when analyzing a very wide frequency range with a very tight
peak, the fitting algorithm can take a lot of iterations (hundreds) to
finally converge on the optimal set of coefficients.  The firmware max
iteration value is set to 40; if the firmware hits the maximum number of
iterations fitting halts and the parameters are returned as-is with a
status code of 5 to indicate that it may not be optimal.

The parameters that are returned *are* the best guess for a fit after
that number of iterations and they aren't necessarily a poor fit.
Instead of considering this a failure on the client side, we now
consider it to have passed and will rely on the R**2 test (like we do
already) to decide if the fit is any good.