tools/modbus: Factor serial.py out of bus.py.
This refactors the existing Bus class into a generic base class and a
serial-port-specific subclass.  The base class does high-level
operations such as Read ID and uses an abstract transact() method to get
the implementation to submit the request and read the response from the
wire.

This also adds a response timeout parameter to the transact() method,
which will be useful for when we implement support for the XtalX Modbus
Adapter board.