p_sensor/xhtism.py: Refactor to take a generic Modbus Bus object.
Previously, the XHTISM class would take an interface, baud rate and
parity and then instantiate a Modbus serial Bus to use.  In the future
we will have different types of busses that can be used in an abstract
way, so change this around so that the client passes a bus down into
XHTISM instead of XHTISM making it from the bottom-up.