xtalx/p_sensor: Factor XHTISS 0.9.1 support into new file; add 0.9.2 support.
XHTISS 0.9.2 adds proper CRC and other error-checking to its protocol.
We now do an initial probe step in make_xhtiss() to try and figure out
which protocol we are talking and then create an 0.9.1-specific object
if it is the old protocol.

In the future we will deprecate support for the old protocol.