p_sensor/xhtiss: Refactor so that the comms protocol is a delegate class.
We don't really want to have two versions of the XHTISS class, since
that makes it hard to do subclassing elsewhere to extend the sensor
capabilities.  The only difference between the versions is the comms
protocol itself, so we factor that out into separate files and bring
back a single XHTISS class that probes for the correct protocol.