z_sensor/xtalx_freq_tool.py: Fix some issues with how the --serial-number argument was handled.
When you have adapters with serial numbers and sensors with serial
numbers, it is a bit confusing and this code was done wrong.  Really, it
should scan all sensors from all adapters to find the sensor with the
matching serial number instead of assuming there is just one adapter
present.